perf

+++ java/src/lib.rs
@@ -43,7 +43,7 @@ fn Java_com_github_hanabi1224_RuAnnoy_NativeMethods_loadIndex_inner(
 ) -> Result<jlong, Box<dyn Error>> {
     let ru_path: String = env.get_string(path)?.into();
     let ru_index_type: IndexType = unsafe { mem::transmute(index_type) };
-    let index = AnnoyIndex::load(dimension, ru_path.as_str(), ru_index_type)?;
+    let index = AnnoyIndex::load(dimension as usize, ru_path.as_str(), ru_index_type)?;
     let ptr = Box::into_raw(Box::new(index));
     return Ok(ptr as jlong);
 }
@@ -83,7 +83,7 @@ ffi_fn! {
     ) -> jfloatArray {
         let index = unsafe { &*(pointer as *const AnnoyIndex) };
         let vector = index.get_item_vector(item_index);
-        let result = env.new_float_array(index.dimension).unwrap();
+        let result = env.new_float_array(index.dimension as i32).unwrap();
         let _ = env.set_float_array_region(result, 0, &vector.as_slice());
         result
     }
@@ -115,7 +115,7 @@ ffi_fn! {
             search_k,
             should_include_distance != 0,
         );
-        let r_id_list: Vec<i64> = r.iter().map(|i| i.id).collect();
+        let r_id_list: Vec<i64> = r.iter().map(|i| i.id as i64).collect();
         let _ = env.set_long_array_region(id_list, 0, &r_id_list.as_slice());
         if should_include_distance != 0 {
             let r_distance_list: Vec<f32> = r.iter().map(|i| i.distance).collect();
@@ -159,7 +159,7 @@ ffi_fn! {
                     search_k,
                     should_include_distance != 0,
                 );
-                let r_id_list: Vec<i64> = r.iter().map(|i| i.id).collect();
+                let r_id_list: Vec<i64> = r.iter().map(|i| i.id as i64).collect();
                 let _ = env.set_long_array_region(id_list, 0, &r_id_list.as_slice());
                 if should_include_distance != 0 {
                     let r_distance_list: Vec<f32> = r.iter().map(|i| i.distance).collect();
