+++ tests/test.rs
@@ -13,7 +13,7 @@ mod tests {
     use std::slice;
     use std::vec::Vec;
 
-    const TEST_INDEX_DIM: i32 = 5;
+    const TEST_INDEX_DIM: usize = 5;
     const TEST_NODE_COUNT: usize = 100;
 
     #[test]
@@ -101,7 +101,7 @@ mod tests {
     fn sanity_tests_inner(
         index_type: IndexType,
         expected_item3_vec: &[f32],
-        expected_id_list: &[i64],
+        expected_id_list: &[u64],
         expected_distance_list: &[f32],
     ) {
         let filepath = format!("tests/index.{}.{}d.ann", index_type, TEST_INDEX_DIM);
@@ -110,14 +110,14 @@ mod tests {
 
         let v0 = index.get_item_vector(0);
         let nearest = index.get_nearest(v0.as_ref(), 5, -1, true);
-        let mut id_list: Vec<i64> = Vec::new();
+        let mut id_list: Vec<u64> = Vec::new();
         let mut distance_list: Vec<f32> = Vec::new();
         for item in &nearest {
             id_list.push(item.id);
             distance_list.push(item.distance);
         }
 
-        assert_eq!(index.degree as usize, TEST_NODE_COUNT);
+        assert_eq!(index.degree, TEST_NODE_COUNT);
         assert_eq!(id_list, expected_id_list);
         assert_eq!(distance_list, expected_distance_list);
         assert_eq!(distance_list.len(), expected_distance_list.len());
@@ -140,7 +140,7 @@ mod tests {
     fn sanity_tests_inner_ffi(
         index_type: IndexType,
         expected_item3_vec: &[f32],
-        expected_id_list: &[i64],
+        expected_id_list: &[u64],
         expected_distance_list: &[f32],
     ) {
         let filepath = format!("tests/index.{}.{}d.ann", index_type, TEST_INDEX_DIM);
@@ -167,7 +167,7 @@ mod tests {
                 let nearest_raw = get_nearest(index, v0_raw, 5, -1, true);
                 let result_count = get_result_count(nearest_raw);
                 let id_list_raw = get_id_list(nearest_raw);
-                let id_list = slice::from_raw_parts(id_list_raw as *mut i64, result_count).to_vec();
+                let id_list = slice::from_raw_parts(id_list_raw as *mut u64, result_count).to_vec();
                 assert_eq!(id_list, expected_id_list);
                 let distance_list_raw = get_distance_list(nearest_raw);
                 let distance_list =
@@ -179,7 +179,7 @@ mod tests {
                 let nearest_raw = get_nearest_to_item(index, 0, 5, -1, true);
                 let result_count = get_result_count(nearest_raw);
                 let id_list_raw = get_id_list(nearest_raw);
-                let id_list = slice::from_raw_parts(id_list_raw as *mut i64, result_count).to_vec();
+                let id_list = slice::from_raw_parts(id_list_raw as *mut u64, result_count).to_vec();
                 assert_eq!(id_list, expected_id_list);
                 let distance_list_raw = get_distance_list(nearest_raw);
                 let distance_list =
@@ -191,7 +191,7 @@ mod tests {
                 let nearest_raw = get_nearest(index, v0_raw, 5, -1, false);
                 let result_count = get_result_count(nearest_raw);
                 let id_list_raw = get_id_list(nearest_raw);
-                let id_list = slice::from_raw_parts(id_list_raw as *mut i64, result_count).to_vec();
+                let id_list = slice::from_raw_parts(id_list_raw as *mut u64, result_count).to_vec();
                 assert_eq!(id_list, expected_id_list);
                 let distance_list_raw = get_distance_list(nearest_raw);
                 let distance_list =
@@ -204,7 +204,7 @@ mod tests {
                 let nearest_raw = get_nearest_to_item(index, 0, 5, -1, false);
                 let result_count = get_result_count(nearest_raw);
                 let id_list_raw = get_id_list(nearest_raw);
-                let id_list = slice::from_raw_parts(id_list_raw as *mut i64, result_count).to_vec();
+                let id_list = slice::from_raw_parts(id_list_raw as *mut u64, result_count).to_vec();
                 assert_eq!(id_list, expected_id_list);
                 let distance_list_raw = get_distance_list(nearest_raw);
                 let distance_list =
